refactor(agent): split policy effect composition
