+++ packages/agent/src/core/policy/effect-merge-rules.ts
@@ -1,0 +1,146 @@
+import type {
+  ApprovalAccumulator,
+  EffectEntry,
+  MergedEffect,
+  MergeResult,
+  PriorityApprovalAccumulator,
+  RetryAccumulator,
+} from "./effect-composition-types";
+import {
+  appendPriorityReason,
+  appendReason,
+  mergeRetry,
+  selectPriorityEffect,
+} from "./effect-merge-accumulators";
+import { appendMergedEffects } from "./effect-merge-output";
+import { assertNever, deepMergeRecords } from "./effect-records";
+import { collectPostConflicts } from "./effect-conflicts";
+
+export function mergeEntries(entries: readonly EffectEntry[]): MergeResult {
+  const merged: MergedEffect[] = [];
+  const postConflicts = collectPostConflicts(entries);
+  const toolFilters = new Map<string, number>();
+  let toolRewrite: { readonly input: Record<string, unknown>; readonly order: number } | undefined;
+  let toolOutputRewrite: MergedEffect | undefined;
+  let toolSkip: MergedEffect | undefined;
+  let toolApproval: ApprovalAccumulator | undefined;
+  let runAbort: MergedEffect | undefined;
+  let continueWithPrompt: MergedEffect | undefined;
+  let retryAfter: RetryAccumulator | undefined;
+  let runReplaceMessages: MergedEffect | undefined;
+  let delegationConstraints:
+    | { readonly constraints: Record<string, unknown>; readonly order: number }
+    | undefined;
+  let delegationApproval: ApprovalAccumulator | undefined;
+  let promptReplace: MergedEffect | undefined;
+  let writebackRewrite: MergedEffect | undefined;
+  let writebackSuppress: PriorityApprovalAccumulator | undefined;
+  let timeout: { readonly timeoutMs: number; readonly order: number } | undefined;
+  let workspaceLock: { readonly required: boolean; readonly order: number } | undefined;
+
+  for (const entry of entries) {
+    const { effect } = entry;
+
+    switch (effect.type) {
+      case "prompt.append_context":
+      case "prompt.inject_message":
+      case "audit.annotate":
+        merged.push({ effect, order: entry.order, priority: entry.priority });
+        break;
+      case "prompt.replace":
+        promptReplace = selectPriorityEffect(promptReplace, entry);
+        break;
+      case "tool.filter":
+        if (!toolFilters.has(effect.toolPattern)) toolFilters.set(effect.toolPattern, entry.order);
+        break;
+      case "tool.rewrite_input":
+        toolRewrite = {
+          input: deepMergeRecords(toolRewrite?.input ?? {}, effect.input),
+          order: Math.min(toolRewrite?.order ?? entry.order, entry.order),
+        };
+        break;
+      case "tool.rewrite_output":
+        toolOutputRewrite = selectPriorityEffect(toolOutputRewrite, entry);
+        break;
+      case "tool.skip_invocation":
+        toolSkip ??= { effect, order: entry.order, priority: entry.priority };
+        break;
+      case "tool.require_approval":
+        toolApproval = appendReason(toolApproval, effect.reason, entry.order);
+        break;
+      case "run.abort":
+        runAbort ??= { effect, order: entry.order, priority: entry.priority };
+        break;
+      case "run.continue_with_prompt":
+        continueWithPrompt = selectPriorityEffect(continueWithPrompt, entry);
+        break;
+      case "run.retry_after":
+        retryAfter = mergeRetry(retryAfter, effect, entry.order);
+        break;
+      case "run.replace_messages":
+        runReplaceMessages = selectPriorityEffect(runReplaceMessages, entry);
+        break;
+      case "delegation.set_constraints":
+        delegationConstraints = {
+          constraints: deepMergeRecords(
+            delegationConstraints?.constraints ?? {},
+            effect.constraints,
+          ),
+          order: Math.min(delegationConstraints?.order ?? entry.order, entry.order),
+        };
+        break;
+      case "delegation.require_approval":
+        delegationApproval = appendReason(delegationApproval, effect.reason, entry.order);
+        break;
+      case "writeback.rewrite":
+        writebackRewrite = selectPriorityEffect(writebackRewrite, entry);
+        break;
+      case "writeback.suppress":
+        writebackSuppress = appendPriorityReason(
+          writebackSuppress,
+          effect.reason,
+          entry.order,
+          entry.priority,
+        );
+        break;
+      case "runtime.set_timeout":
+        timeout = {
+          timeoutMs: Math.min(timeout?.timeoutMs ?? effect.timeoutMs, effect.timeoutMs),
+          order: Math.min(timeout?.order ?? entry.order, entry.order),
+        };
+        break;
+      case "runtime.workspace_lock":
+        workspaceLock = {
+          required: (workspaceLock?.required ?? false) || effect.required,
+          order: Math.min(workspaceLock?.order ?? entry.order, entry.order),
+        };
+        break;
+      default:
+        assertNever(effect);
+    }
+  }
+
+  appendMergedEffects(merged, {
+    promptReplace,
+    toolFilters,
+    toolRewrite,
+    toolOutputRewrite,
+    toolSkip,
+    toolApproval,
+    runAbort,
+    continueWithPrompt,
+    retryAfter,
+    runReplaceMessages,
+    delegationConstraints,
+    delegationApproval,
+    writebackRewrite,
+    writebackSuppress,
+    timeout,
+    workspaceLock,
+  });
+
+  return {
+    effects: merged.sort((left, right) => left.order - right.order).map(({ effect }) => effect),
+    postConflicts,
+  };
+}
